+++ CHANGELOG.md
@@ -32,6 +32,7 @@
 - Episode title, duration, number and pubDate are not overwritten when fetching metadata if they are not empty
 - Refactored console ResaveController to add new options and support core 'set' option
 - Changed podcast RSS URL to use site handle instead of site Id
+- Podcast episode settings can be set per site
 - Fixed route for episodes
 - Fixed a bug where changing lightswitch inputs on the sidebar of the episode edit page, creates unwanted drafts
 - Fixed permission issues where non-admin users were not able to access some plugin actions